fix: move hooks to service dir, center delete title
